Does not self allocate IDs in Beam by default.
Per b/250948425, it is dangerous to implicitly allow all Beam pipelines
to create buildables by self allocating the IDs. This change makes it so
that one has to explicitly request self allocation in Beam.

The boolean is added to the pipeline option so that it can be passed to
the beam worker initializer that controls the behavior of the JVM on
each worker. We did not expose it in the metadata.json file because we did
not want to be able to specify the override at run time, due to the risk. As
shown in RdePipeline.java, we instead explicitly hard-code the option.